"Threshold by sum" pixel filtering incorporated
Example Code has been modified such that it can apply a secondary pixel mask (based on intensity threshold applied in SPCImage) in addition to the ROI mask (from mask TIFF file) when calling the regionprops_omi function. This allows the user to threshold out low-count pixels in SPCImage that produce bad fits and can skew summary results. A threshold by sum with a minumum value of 250 photons in SPCImage is recommended.